requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
idna==3.3
lxml==4.8.0
pyquery==1.4.3
python-dateutil==2.8.2
python-decouple==3.6
python-metallum==1.1.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-cache==0.9.3
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
url-normalize==1.4.3
urllib3==1.26.8
